refactor: decouple codeagent-wrapper with AgentBackend plugin architecture
- Introduce AgentBackend Protocol in vibe3.agents.base
- Extract CodeagentBackend implementation to vibe3.agents.backends.codeagent
- Migrate orchestrators and usecases from vibe3.services to vibe3.agents
- Update command layer and other services to use the new agents module
- Simplify architecture by removing direct binary dependency in core services

Resolves: #394
