+++ src/vibe3/agents/__init__.py
@@ -1,0 +1,3 @@
+from vibe3.agents.base import AgentBackend
+
+__all__ = ["AgentBackend"]
