+++ src/vibe3/agents/review_agent.py
@@ -1,4 +1,7 @@
-"""Usecase layer for review command orchestration."""
+"""Usecase layer for review command orchestration.
+
+Migrated from vibe3.services.review_usecase.
+"""
 
 from dataclasses import dataclass
 from typing import Any, Callable, cast
@@ -6,21 +9,21 @@ from typing import Any, Callable, cast
 import typer
 from loguru import logger
 
+from vibe3.agents.review_parser import ParsedReview, parse_codex_review
+from vibe3.agents.runner import (
+    CodeagentExecutionService,
+    create_codeagent_command,
+)
 from vibe3.clients.github_client import GitHubClient
 from vibe3.clients.github_issues_ops import parse_linked_issues
 from vibe3.config.settings import VibeConfig
 from vibe3.models.orchestration import IssueState
 from vibe3.models.review import ReviewRequest, ReviewScope
 from vibe3.models.snapshot import StructureDiff
-from vibe3.services.codeagent_execution_service import (
-    CodeagentExecutionService,
-    create_codeagent_command,
-)
 from vibe3.services.context_builder import make_review_context_builder
 from vibe3.services.flow_service import FlowService
 from vibe3.services.inspect_output_adapter import changed_symbols
 from vibe3.services.label_service import LabelService
-from vibe3.services.review_parser import ParsedReview, parse_codex_review
 from vibe3.services.review_pipeline_helpers import build_snapshot_diff, run_inspect_json
