+++ src/vibe3/commands/plan.py
@@ -5,6 +5,11 @@ from typing import Annotated, Optional
 
 import typer
 
+from vibe3.agents.plan_agent import PlanUsecase
+from vibe3.agents.runner import (
+    CodeagentExecutionService,
+    create_codeagent_command,
+)
 from vibe3.clients.github_client import GitHubClient
 from vibe3.commands.command_options import (
     _AGENT_OPT,
@@ -19,16 +24,11 @@ from vibe3.commands.command_options import (
 from vibe3.config.settings import VibeConfig
 from vibe3.models.orchestration import IssueState
 from vibe3.models.plan import PlanRequest
-from vibe3.services.codeagent_execution_service import (
-    CodeagentExecutionService,
-    create_codeagent_command,
-)
 from vibe3.services.flow_service import FlowService
 from vibe3.services.label_service import LabelService
 from vibe3.services.plan_context_builder import (
     make_plan_context_builder,
 )
-from vibe3.services.plan_usecase import PlanUsecase
 from vibe3.services.spec_ref_service import SpecRefService
 from vibe3.utils.trace import enable_trace
