+++ src/vibe3/orchestra/master.py
@@ -8,7 +8,7 @@ from pathlib import Path
 
 from loguru import logger
 
-from vibe3.services.review_runner import (
+from vibe3.agents.review_runner import (
     DEFAULT_WRAPPER_PATH,
     AgentOptions,
 )
